Fjerner ubrukt komponent
